Input Locking and Directory Tree Privacy
The directory tree is now saved to a directory known as "private" in the root directory of Z to easily keep sensitive information safe.

Inputs can be locked so that they are not cleared when an entry is submitted. This is helpful for creating a large number of entries with similar attributes.

Started working on a feature that would allow different components of Z to be allocated to their own respective windows. However, this feature is currently experiencing errors. In the future, I intend to be able to detach these components from one another.